Compiles (with many warnings) but segfaults when trying to run the crypto methods. Also some weirdness going on with cmdline options.

+++ crypt.h
@@ -1,6 +1,9 @@
 #ifndef _GUARD_CRYPT
 #define _GUARD_CRYPT
 
+#include <stdint.h>
+#include "primitives.h"
+
 typedef enum {
 	ENCRYPT,
 	DECRYPT
@@ -24,15 +27,15 @@ typedef enum {
 
 void rc5_ecb_encrypt16(data16 ptext, data16 ctext, uint16_t *ksched, size_t rounds, padmode_t padmode);
 void rc5_ecb_encrypt32(data32 ptext, data32 ctext, uint32_t *ksched, size_t rounds, padmode_t padmode);
-void rc5_ecb_encrypt64(data64 ptext, data64 ctext, uint32_t *ksched, size_t rounds, padmode_t padmode);
+void rc5_ecb_encrypt64(data64 ptext, data64 ctext, uint64_t *ksched, size_t rounds, padmode_t padmode);
 void rc5_ecb_decrypt16(data16 ctext, data16 ptext, uint16_t *ksched, size_t rounds, padmode_t padmode);
 void rc5_ecb_decrypt32(data32 ctext, data32 ptext, uint32_t *ksched, size_t rounds, padmode_t padmode);
 void rc5_ecb_decrypt64(data64 ctext, data64 ptext, uint64_t *ksched, size_t rounds, padmode_t padmode);
 void rc5_cbc_encrypt16(data16 ptext, data16 ctext, uint16_t *ksched, size_t rounds, padmode_t padmode);
 void rc5_cbc_encrypt32(data32 ptext, data32 ctext, uint32_t *ksched, size_t rounds, padmode_t padmode);
-void rc5_cbc_encrypt64(data64 ptext, data64 ctext, uint32_t *ksched, size_t rounds, padmode_t padmode);
+void rc5_cbc_encrypt64(data64 ptext, data64 ctext, uint64_t *ksched, size_t rounds, padmode_t padmode);
 void rc5_cbc_decrypt16(data16 ctext, data16 ptext, uint16_t *ksched, size_t rounds, padmode_t padmode);
 void rc5_cbc_decrypt32(data32 ctext, data32 ptext, uint32_t *ksched, size_t rounds, padmode_t padmode);
-void rc5_cbc_decrypt64(data64 ctext, data64 ptext, uint32_t *ksched, size_t rounds, padmode_t padmode);
+void rc5_cbc_decrypt64(data64 ctext, data64 ptext, uint64_t *ksched, size_t rounds, padmode_t padmode);
 
 #endif
